Dynamically generating fault based on node selector
Since we need to selectively crash pods depending on their role (leader or not), now we generate multiple faults for pod failure, and change the percentage of pod getting crashed based on its role

For cass-operator, the leader label selection is always written first in the fault injection config. So for the first label, chactos always crash all of the pods (i.e. crashing all leaders), then we crash some of the non-leader nodes